+++ daemons/dss-scalability-test/.chalice/config.json
@@ -1,13 +1,14 @@
 {
   "version": "2.0",
   "app_name": "dss-scalability-test",
+  "environment_variables": {
+    "HOME": "/tmp",
+    "HOME_comment": "HOME variable is necessary in order to create templfile/tempdirectories in Lambda function"
+  },
   "stages": {
     "dev": {
       "api_gateway_stage": "api",
-      "environment_variables": {
-        "HOME": "/tmp",
-        "HOME_comment": "HOME variable is necessary in order to create templfile/tempdirectories in Lambda function"
-      }
+      "environment_variables": {}
     }
   },
   "lambda_timeout": 300,
