+++ tests/scalability/scale_test_runner.py
@@ -33,7 +33,6 @@ class ScaleTestRunner:
             self.loading = self.loading[:cnt] + '#' + self.loading[cnt + 1:]
             print('\r%s Sending at %3d percent!' % (self.loading, (cnt + 1) * 100 / self.duration_seconds), end='')
         print("\nDone")
-        self.sns_client.stop()
 
     def generate_load_rps(self):
         start = datetime.datetime.now()
